writefil.c: support diagonals to the right and to the left
The Postscript code has been changed to draw diagonals to the right if
the argument passed to draw_diagonals has bit 0 set. It'll draw
diagonals to the left if the argument has bit 1 set. If both bits are
set, both diagonals are being drawn.

The argument of the -i switch is used as the value passed into
draw_diagonals is, i.e.  if you call 'pcal -i 1 ...' it'll draw
diagonals to the right, 'pcal -i 2 ...' will draw diagonals to the left
and 'pcal -i 3 ...' will draw both.